Exception handling.
Add the empty.rfa file to ensure there's always something to open.
Don't try and load results if there are no results to load.
Make sure the run button has a command parameter.

+++ src/RevitTestFrameworkRunner/ViewModel.cs
@@ -183,6 +183,12 @@ namespace RevitTestFrameworkRunner
 
         private void Run(object parameter)
         {
+            if (string.IsNullOrEmpty(Program._results))
+            {
+                MessageBox.Show("Please select an output path for the results.");
+                return;
+            }
+
             if (File.Exists(Program._results) && !Program._concat)
             {
                 File.Delete(Program._results);
